8256955: Move includes of events.hpp out of header files
